authorization and pulling form data are now done in a single component
